+++ src/utils/generateLine.ts
@@ -1,0 +1,5 @@
+import { insertText } from "./insertText";
+
+export function generateLine(): void {
+	insertText();
+}