+++ src/utils/generateParagraph.ts
@@ -1,0 +1,6 @@
+import { insertText } from "./insertText";
+import { randomInt } from "./randomInt";
+
+export function generateParagraph(): void {
+	insertText(randomInt(4) + 10);
+}