+++ src/utils/randomInt.ts
@@ -1,0 +1,3 @@
+export function randomInt(max: number): number {
+	return Math.floor(Math.random() * Math.floor(max));
+}